Filled in constrCheck() and changed varHuer()
I don't think the constrCheck() is right... I wasn't sure how to get the values to check against the constraints '-_- and for the varHuer() it kept outputting the first Variable wrong. I made changes and commented out the old stuff. I'm not sure if what I added works all the way. I tested it and it will output F: 1, then F:2. But I think there's a problem with constrCheck() so it's not continuing and it's not outputting any other variables past F....